Resolving sign mismatch & implicit truncation warnings

+++ src/sqlgen/mysql/Connection.cpp
@@ -112,7 +112,7 @@ typename Connection::ConnPtr Connection::make_conn(
   const auto res = mysql_real_connect(
       shared_ptr.get(), _credentials.host.c_str(), _credentials.user.c_str(),
       _credentials.password.c_str(), _credentials.dbname.c_str(),
-      _credentials.port, _credentials.unix_socket.c_str(),
+      static_cast<uint>(_credentials.port), _credentials.unix_socket.c_str(),
       CLIENT_MULTI_STATEMENTS);
 
   if (!res) {
@@ -140,7 +140,7 @@ Result<Ref<Iterator>> Connection::read_impl(
   const auto sql =
       _query.visit([](const auto& _q) { return mysql::to_sql_impl(_q); });
   const auto err =
-      mysql_real_query(conn_.get(), sql.c_str(), static_cast<int>(sql.size()));
+      mysql_real_query(conn_.get(), sql.c_str(), sql.size());
   if (err) {
     return make_error(conn_);
   }
